todo: 1、图片验证码识别(已完成，接入第三方接口） 2、修复开启代理导致的卡顿(已完成） 3、省城市和街道地址的选择 4、hook接口修改手机设备信息瞒过QQ 5、找商品准确率不高
finished：1,2

完成一整套流程接口:VPN service需要重启杀死，取账号的接口有问题

+++ accessibility_service/src/main/java/com/accessibility/service/data/TaskBean.java
@@ -63,7 +63,7 @@ public class TaskBean {
          * pay_account : {"id":"","type":"","username":"","pwd":""}
          */
 
-        private long task_id;
+        private int task_id;
         private int task_type;
         private int task_status;
         private int task_amount;
@@ -82,11 +82,11 @@ public class TaskBean {
         private StayTimeBean stay_time;
         private PayAccountBean pay_account;
 
-        public long getTask_id() {
+        public int getTask_id() {
             return task_id;
         }
 
-        public void setTask_id(long task_id) {
+        public void setTask_id(int task_id) {
             this.task_id = task_id;
         }
